requirements.txt
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
psutil==5.8.0
google-cloud-secret-manager==2.0.0
google-cloud-pubsub==2.2.0
Flask-Mail==0.9.1